Move files out from components folder
