+++ src/components/Career.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import styled from 'styled-components';
-import { Section, SectionBody, SectionTitle } from '../Section';
-import careerSummary from '../../globals/data/careerSummary';
+import { Section, SectionBody, SectionTitle } from './Section';
+import careerSummary from '../globals/data/careerSummary';
 
 const CareerParagraph = styled.p`
   margin: 0.5rem 0;
