+++ src/components/Jobs.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import styled from 'styled-components';
-import experiences from '../../globals/data/experiences';
-import { generateTimeRange } from '../../utils/date.utils';
-import { Section, SectionBody, SectionTitle } from '../Section';
+import experiences from '../globals/data/experiences';
+import { generateTimeRange } from '../utils/date.utils';
+import { Section, SectionBody, SectionTitle } from './Section';
 
 export const Job = styled.div`
   margin-top: 1rem;
